requirements.txt
proton-core @ git+https://github.com/ProtonVPN/python-proton-core.git
